+++ shared/featureconfig/flags.go
@@ -129,6 +129,10 @@ var (
 		Name:  "enable-initial-sync-queue",
 		Usage: "Enables concurrent fetching and processing of blocks on initial sync.",
 	}
+	enableFieldTrie = cli.BoolFlag{
+		Name:  "enable-state-field-trie",
+		Usage: "Enables the usage of state field tries to compute the state root",
+	}
 )
 
 // Deprecated flags list.
@@ -308,6 +312,7 @@ var BeaconChainFlags = append(deprecatedFlags, []cli.Flag{
 	broadcastSlashingFlag,
 	newStateMgmt,
 	enableInitSyncQueue,
+	enableFieldTrie,
 }...)
 
 // E2EBeaconChainFlags contains a list of the beacon chain feature flags to be tested in E2E.
